Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-scoping/slotted-slot-expected.txt

This text should be green.

FAIL Check that ::slotted does not match slot elements assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
